+++ src/app/static/src/tests/adr/mutations.test.ts
@@ -63,5 +63,4 @@ describe("ADR mutations", () => {
         mutations[ADRMutation.SetUploadFiles](state, {payload});
         expect(state.uploadFiles).toBe(payload);
     });
-
 });