+++ NEWS.md
@@ -1,6 +1,10 @@
-£ hint 1.27.0
+# hint 1.28.0
 
 * Avoid re-uploading unchanged files to ADR
+
+# hint 1.27.0
+
+* Invoke ADR uploads with actions
 
 # hint 1.26.0